Fixing a corrupt data in job.runRusage
When the get_job_info exits, it destroys the internal pointers (intp) in
the python Swig object that links to the LSF C API, which means they
point randomly to chunks of memory and have unknown random amounts when
you try to look at them later. One needs to do a deep copy on the value
itself while the object is still valid, and pass that result back
upstream.